Add support for saving global snapshot into file
- Add possibility to save symbol tree structure
parsed from pdb file into json so that symbols
can be loaded from there later on even if pdb
file does not exist. Currently the names of pointed
function for function pointers are lost but restoring
the pointer for snapshot still works
- Add suppport for saving snapshot of global symbols
into a file and restoring the values from there on
another execution. The snapshot requires either
the json structure for symbols or the original pdb
file.