Improve signal deletion
Only mark signals as deleted and handle the actual deletion
in updateSavedSettings. This way the windows do not need
to touch the saved json settings as much and deletion is handled
in a single place. Also signals deleted from scalar/vector window
are now properly deleted from the plots as well