Add checkbox for searching symbols recursively
Sometimes it can be difficult to find the name of the parent
symbol so add option for searching directly all members of
the symbols directly. This can be quite slow however and can give
too many results so it is disabled by default. The option is not saved
to json file because it is probably not usually used