Fix filtering signals with intermediate group match
Signals within subgroups should be shown if the
upper level parent group matches instead of only
the direct parent.

Currently all subgroups get expanded if the
parent group matches which is not ideal. Only
the direct match group should be expanded
but the subgroups should require clicking

Also fix that the match to group is not checked
for every signal