Change loading symbols to be a singleton
SymInitialize if called multiple times and looking up the information a
second time is not useful when it has been looked up for the first object
already anyway. Change the object using pdb file to be a singleton so
that it will be shared automatically.

When loading symbols from json, it can fail so there can be multiple
objects and it does not use the dbghelp library in that case.

Changing saving symbol info to json to be a separate function so
that it is clearer which method is used for loading symbol info.