Always call SymCleanup after SymInitialize
If two dlls within the same process call SymInitialize, the
second call will fail because the first one did not call SymCleanup.
There is no easy way to detect if SymInitialize has been called
already so the only sane way seems to be that the symbol handler
cannot be left open for the function pointer lookup.

Always reinitialize symbol handler when searching function
pointers but cache the addresses->symbol because otherwise
the lookup is very slow and makes the gui sluggish if there
are multiple function pointers in the search window.