feat: Add support for nested docks in arbitrary order
The ImGui::DockSpace states "DockSpace() needs to be submitted _before_ any window they can
host" so the windows need to be created in such order that the window to which the dockspace
is tried to dock is created first. This is done by moving the dragged window to be the last
dockspace to be created so it can be docked to any other dockspace.