feat: Add method for displaying message to main menu bar
Making a separate console window is not a good
option because all the prime estate is reserved for
the plots so display the message as just one line
on the main menu bar and more messages are
visible when hovering. Would like a better
solution but maybe this is sufficient.